feat: comment

+++ src/main/kotlin/com/readnumber/dailyonestep/common/dto/MultipleResourceResponseDto.kt
@@ -1,13 +1,13 @@
 package com.readnumber.dailyonestep.common.dto
 
 class MultipleResourceResponseDto<T>(
-    val totalCount: Long,
+    val totalCount: Int,
     val resources: List<T>
 ) {
     companion object {
         fun <T> from(resources: List<T>): MultipleResourceResponseDto<T> {
             return MultipleResourceResponseDto(
-                totalCount = resources.size.toLong(),
+                totalCount = resources.size,
                 resources = resources
             )
         }
