+++ src/main/kotlin/com/readnumber/dailyonestep/post/dto/response/MultiplePostWrapperDto.kt
@@ -1,8 +1,7 @@
 package com.readnumber.dailyonestep.post.dto.response
 
-import io.swagger.v3.oas.annotations.media.Schema
 
 class MultiplePostWrapperDto(
-    val totalCount: Long,
+    val totalCount: Int,
     val posts: List<PostDto>
 )
